Eject tools/types.ts from the daemon core import cycle
tools/types.ts is imported by ~197 files but was itself trapped in the giant
daemon-core SCC by just two type-only back-edges into heavy core modules. Cut
both and clean up its re-export surface.

- SecretPromptResult (+ SecretDelivery) moved out of the heavy
  permissions/secret-prompter.ts into a new leaf permissions/secret-prompt-types.ts;
  the prompter and all type-only consumers import from the leaf.
- TrustClass already lives in the leaf runtime/trust-class.ts; tools/types.ts
  imported it via the heavy runtime/actor-trust-resolver.ts (which re-exports
  it). Repointed to the leaf.

With both edges cut, tools/types.ts has no import into the core and drops out
of the SCC.

Also removed the compat re-export block that mirrored ./tool-types.js through
tools/types.ts (the canonical declarations already live in the neutral leaf)
and repointed the consumers of the re-exported names (RiskLevel,
ExecutionTarget, ProxyEnvVars, ProxyApprovalCallback/Request,
ToolPermissionDenied/PromptEvent) to import from tools/tool-types.js directly.

Core SCC: 479 -> 367 files (-112). tools/types.ts is out of every cycle. Two
smaller pre-existing sub-cycles (config/skills <-> skills/tool-manifest and
daemon/message-protocol <-> message-types/subagents <-> subagent/types) are now
visible as their own isolated SCCs — they were masked inside the blob, not
introduced here.
