+++ assistant/src/__tests__/secret-response-routing.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, expect, mock, test } from "bun:test";
 
 import type { SecretRequestEvent } from "../api/events/secret-request.js";
 import type { ServerMessage } from "../daemon/message-protocol.js";
-import type { SecretPromptResult } from "../permissions/secret-prompter.js";
+import type { SecretPromptResult } from "../permissions/secret-prompt-types.js";
 
 let broadcastedMessages: ServerMessage[] = [];
 mock.module("../runtime/assistant-event-hub.js", () => ({
