+++ assistant/src/daemon/handlers/shared.ts
@@ -9,7 +9,7 @@ import type {
 import { ConfirmationDecisionSchema } from "../../api/responses/conversation-message.js";
 import { getConfig } from "../../config/loader.js";
 import type { LLMCallSite, Speed } from "../../config/schemas/llm.js";
-import type { SecretPromptResult } from "../../permissions/secret-prompter.js";
+import type { SecretPromptResult } from "../../permissions/secret-prompt-types.js";
 import { isPlaceholderSentinelText } from "../../providers/placeholder-sentinels.js";
 import { broadcastMessage } from "../../runtime/assistant-event-hub.js";
 import type { AuthContext } from "../../runtime/auth/types.js";
