+++ assistant/src/tools/skills/skill-script-runner.ts
@@ -2,11 +2,8 @@ import { basename, join, resolve } from "node:path";
 
 import { bundledToolRegistry } from "../../config/bundled-tool-registry.js";
 import { computeSkillVersionHash } from "../../skills/version-hash.js";
-import type {
-  ExecutionTarget,
-  ToolContext,
-  ToolExecutionResult,
-} from "../types.js";
+import type { ExecutionTarget } from "../tool-types.js";
+import type { ToolContext, ToolExecutionResult } from "../types.js";
 import { runSkillToolScriptSandbox } from "./sandbox-runner.js";
 import type { SkillToolScript } from "./script-contract.js";
